+++ docs/.vitepress/config.ts
@@ -27,6 +27,12 @@ export default defineConfig({
     sidebar: {
       '/guide/': [
         {
+          text: 'Rationale',
+          items: [
+            { text: 'Why the library exists', link: '/guide/rationale' },
+          ]
+        },
+        {
           text: 'Getting Started',
           items: [
             { text: 'Installation', link: '/guide/getting-started#installation' },
@@ -51,13 +57,16 @@ export default defineConfig({
             { text: 'validation', link: '/examples/validators' },
             { text: 'groups', link: '/examples/groups' },
             { text: 'density', link: '/examples/density' },
+            { text: 'configuration', link: '/examples/configuration' },
             { text: 'responsive options', link: '/examples/responsive-render-options' },
+            { text: 'localisation', link: '/examples/localisation' },
             { text: 'df-actions', link: '/examples/df-actions' },
             { text: 'df-checkbox', link: '/examples/df-checkbox' },
             { text: 'df-color', link: '/examples/df-color' },
             { text: 'df-date-time', link: '/examples/df-datetime' },
             { text: 'df-file', link: '/examples/df-file' },
             { text: 'df-input', link: '/examples/df-input' },
+            { text: 'df-input-hint', link: '/examples/df-input-hint' },
             { text: 'df-rtf-editor', link: '/examples/df-rtf-editor' },
             { text: 'df-select', link: '/examples/df-select' },
             { text: 'df-text-area', link: '/examples/df-text-area' },
@@ -66,7 +75,7 @@ export default defineConfig({
       ]
     },
     socialLinks: [
-      { icon: 'github', link: 'https://github.com/velis74/dynamicforms-vuetify-inputs' }
+      { icon: 'github', link: 'https://github.com/dynamicforms/vuetify-inputs' }
     ],
     footer: {
       message: 'Released under the MIT License.',